Refactor subscription constants to one file
